changes suggested by Peli
