+++ spectool/jdtest.ts
@@ -1,7 +1,8 @@
 /// <reference path="jdspec.d.ts" />
 /// <reference path="jdtest.d.ts" />
 
-import { parseIntFloat, getRegister } from "./jdutils";
+import { parseIntFloat, getRegister, packetsToRegisters } from "./jdutils";
+import { camelize, capitalize } from "./jdspec"
 
 const exprTokens = ["(", ")", "+", "-", "<", ">", "<=", ">=", "==", "/", 
 "*", "&&", "||", ",", "[", "]", "%", ">>", "<<", "&", "|",
@@ -209,3 +210,88 @@ export function parseSpecificationTestMarkdownToJSON(filecontent: string, spec: 
     }
 }
 
+function gatherLocals(serviceTest: jdtest.ServiceTest) {
+    let locals: string[] = []
+    serviceTest.tests.forEach(t => {
+        t.letVariables.forEach(l => {
+            if (locals.indexOf(l) < 0)
+                locals.push(l);
+        })
+    })
+    return locals;
+}
+
+// create a file to check syntax of expressions, if this file passes tsc
+// then toTestMonitorClient code will be fine
+
+export function testSyntax(spec: jdspec.ServiceSpec, serviceTest: jdtest.ServiceTest) {
+    const { shortId, name, camelName, packets } = spec;
+    const Reading = 0x101;
+
+    const registers = packetsToRegisters(packets);
+    const regs = registers.filter(r => !!r);
+    const locals = gatherLocals(serviceTest);
+
+    return `namespace test_syntax {
+${regs.map(reg => `
+        let ${reg.name} = 0;`).join("")}            
+${locals.map(l => `
+        let ${l} = 0;`).join("")}      
+${serviceTest.tests.map(t => t.commands.map(c => `
+        ${c.expr.map(exprToString).join(" ")};
+`).join("")).join("")}
+    }`;
+    
+    function exprToString(token: jdtest.ServiceTestToken) {
+        return (token.js ? token.js : "") + (token?.const ? token.const.toString() : "") + (token.id ? token.id : "");
+    }
+}
+
+// create a test monitor as a helper for manual tests
+
+// use reg.subscribe(CHANGE, () =>. ....) to listen for changes
+// await regchange(reg, v => v > 5), but need a hook for cancelling
+
+export function toTestMonitor(spec: jdspec.ServiceSpec, serviceTest: jdtest.ServiceTest) {
+    const { camelName, packets } = spec;
+
+    const registers = packetsToRegisters(packets);
+    const regs = registers.filter(r => !!r);
+    const locals = gatherLocals(serviceTest);
+    const capName = capitalize(camelName);
+    const className = `${capName}TestClient`
+
+    return `
+    import { JDService } from "../../../src/jdom/service"
+    import { JDServiceClient } from "../../../src/jdom/serviceclient"
+    import { JDRegister } from "../../../src/jdom/register"
+    import { useRegisterUnpackedValue } from "../../../docs/src/jacdac/useRegisterValue";
+    import { ${capName}Reg } from "../specconstants";
+
+    export class ${className} extends JDServiceClient {
+${regs.map(reg => `
+        private ${camelize(reg.name)}Reg: JDRegister
+        private ${camelize(reg.name)}: number;
+        `).join("")}            
+${locals.map(l => `
+        private ${l}: number = 0;`).join("")}  
+    
+        constructor(service: JDService) {
+            super(service);
+${regs.map(reg => `
+            this.${camelize(reg.name)}Reg = service.register(${capName}Reg.${capitalize(camelize(reg.name))});
+            `).join("")}
+        }
+${serviceTest.tests.map((t,t_index) => t.commands.map((c,c_index) => `
+        public test${t_index}_cmd${c_index}() {
+            let expr = ${c.expr.map(exprToString).join(" ")};
+            ${(c.kind === "let") ? `this.${c.lhs} = expr;` : `return expr`}
+        };
+        `
+        ).join("")).join("")}
+    }`;
+
+    function exprToString(token: jdtest.ServiceTestToken) {
+        return (token.js ? token.js : "") + (token?.const ? token.const.toString() : "") + (token.id ? `this.${camelize(token.id)}` : "");
+    }
+}
